requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==7.1.2
distro==1.9.0
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==8.2
idna==3.11
jiter==0.12.0
numpy==2.3.4
openai==2.7.2
playwright==1.55.0
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
python-dotenv==1.0.1

boto3==1.35.90
botocore==1.35.90

SQLAlchemy==2.0.36
setuptools==80.9.0
sniffio==1.3.1
starlette==0.49.3
tabulate==0.8.7
tinydb==4.2.0
tqdm==4.51.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wheel==0.45.1
psycopg2-binary==2.9.10
pgvector==0.2.5
